# apps/api/swag/openapi3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vyzorix Update Server API",
    "version": "0.0.01",
    "description": "Go-generated OpenAPI contract from handler annotations"
  },
  "paths": {
...
    "/commands/{imei}/history": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "commands"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
    "/config-versions/{resource}": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "config-versions",
          "config-versions"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
    "/config-versions/{resource}/{version}": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "config-versions",
          "config-versions"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
    "/config-versions/{resource}/{version}/restore": {
      "post": {
        "summary": "",
        "description": "",
        "tags": [
          "config-versions",
          "config-versions"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
...
    "/dashboard/stats": {
      "get": {
        "summary": "",
        "description": "",
        "tags": [
          "dashboard",
          "dashboard"
        ],
        "parameters": [
          {
            "name": "X-Organization-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Organization ID"
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}